requirements.txt
PyQt6>=6.6.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.1
langchain-ollama>=0.0.1
pynput>=1.7.6
pymupdf>=1.23.0
python-docx>=1.0.0
python-pptx>=0.6.23
pandas>=2.1.0
openpyxl>=3.1.2
